pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rose"
version = "0.1.0"
description = "ROSE - Reflectometry Open Science Engine: reflectometry experiment planner"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "BSD-3-Clause" }
authors = [{ name = "Neutrons AI Team" }]
keywords = [
  "neutron",
  "reflectivity",
  "thin-film",
  "analysis",
  "science",
  "automation",
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: BSD License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Scientific/Engineering",
  "Topic :: Scientific/Engineering :: Physics",
]

# Core dependencies
dependencies = [
  "numpy>=1.24.0",
  "pandas>=2.0.0",
  "scipy>=1.11.0",
  "pyyaml>=6.0.0",
  "refl1d>=1.0.0",
]

[project.optional-dependencies]
# Development tools
dev = [
  "pytest>=7.4.0",
  "pytest-cov>=4.1.0",
  "ruff>=0.1.0",
  "black>=23.0.0",
  "mypy>=1.5.0",
  "pre-commit>=3.4.0",
]

# Web application dependencies
web = ["flask>=3.0.0", "jinja2>=3.1.0", "bootstrap-flask>=2.3.0"]

# API/MCP server dependencies
api = ["fastapi>=0.104.0", "uvicorn[standard]>=0.24.0", "pydantic>=2.4.0"]

# CLI dependencies
cli = ["click>=8.1.0", "tqdm>=4.66.0", "python-dotenv>=1.0.0"]

# Data science extras
science = [
  "scipy>=1.11.0",
  "matplotlib>=3.8.0",
  "plotly>=5.17.0",
  "scikit-learn>=1.3.0",
]

# LLM model generation (use-case 2)
llm = [
  "langchain>=0.3.0",
  "langchain-openai>=0.2.0",
  "pydantic>=2.4.0",
  "python-dotenv>=1.0.0",
  "globus-sdk>=3.0.0",
]

# Documentation
docs = ["sphinx>=7.2.0", "sphinx-rtd-theme>=1.3.0", "myst-parser>=2.0.0"]

# All optional dependencies
all = ["rose[dev,web,api,cli,science,llm,docs]"]

[project.urls]
Homepage = "https://github.com/neutrons-ai/rose"
Documentation = "https://github.com/neutrons-ai/rose#readme"
Repository = "https://github.com/neutrons-ai/rose"
Issues = "https://github.com/neutrons-ai/rose/issues"
AuRE-v1 = "https://github.com/neutrons-ai/aure"

[project.scripts]
rose = "rose.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
rose = ["py.typed"]

# Pytest configuration
[tool.pytest.ini_options]
minversion = "7.0"
pythonpath = ["src"]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
  "-ra",
  "--strict-markers",
  "--strict-config",
  "--cov=src/rose",
  "--cov-report=term-missing",
  "--cov-report=html",
]
markers = [
  "slow: marks tests as slow (deselect with '-m \"not slow\"')",
  "integration: marks tests as integration tests",
  "unit: marks tests as unit tests",
]

# Coverage configuration
[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/test_*.py"]

[tool.coverage.report]
exclude_lines = [
  "pragma: no cover",
  "def __repr__",
  "raise AssertionError",
  "raise NotImplementedError",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
  "@abstractmethod",
]

# Ruff configuration (linting and formatting)
[tool.ruff]
line-length = 88
target-version = "py39"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP", "ARG", "SIM"]
ignore = ["E501", "B008", "C901", "W191"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**/*.py" = ["ARG", "S101"]

[tool.ruff.lint.isort]
known-first-party = ["rose"]

# Black configuration (code formatting)
[tool.black]
line-length = 88
target-version = ["py39", "py310", "py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
    \.git
  | \.venv
  | build
  | dist
)/
'''

# MyPy configuration (type checking)
[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
